# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fnashha Home Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /governorates:
    get:
      operationId: listGovernorates
      tags: [governorates]
      summary: List governorates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Governorate"
    post:
      operationId: createGovernorate
      tags: [governorates]
      summary: Create governorate (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"
  /governorates/{id}:
    patch:
      operationId: updateGovernorate
      tags: [governorates]
      summary: Update governorate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernorateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Governorate"
    delete:
      operationId: deleteGovernorate
      tags: [governorates]
      summary: Delete governorate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AREAS ───────────────────────────────────────────────────────────────────
  /areas:
    get:
      operationId: listAreas
      tags: [areas]
      summary: List areas
      parameters:
        - name: governorateId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Area"
    post:
      operationId: createArea
      tags: [areas]
      summary: Create area (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
  /areas/{id}:
    patch:
      operationId: updateArea
      tags: [areas]
      summary: Update area (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AreaUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
    delete:
      operationId: deleteArea
      tags: [areas]
      summary: Delete area (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── REQUESTS ────────────────────────────────────────────────────────────────
...
  /requests/{requestId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── RATINGS ─────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # ─── CMS ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, mobile, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, technician, admin, super_admin]
        status:
          type: string
          enum: [active, pending, suspended, banned, rejected]
        profileImage:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
        technicianProfile:
          $ref: "#/components/schemas/TechnicianProfile"
...
    TechnicianProfile:
      type: object
      required: [id, userId, nationalId, approvalStatus]
      properties:
        id:
          type: integer
        userId:
          type: integer
        nationalId:
          type: string
        personalPhoto:
          type: ["string", "null"]
        nationalIdFront:
          type: ["string", "null"]
        nationalIdBack:
          type: ["string", "null"]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        pointsBalance:
          type: integer
        averageRating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        areas:
          type: array
          items:
            $ref: "#/components/schemas/Area"
        primaryArea:
          $ref: "#/components/schemas/Area"
...
    Service:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
...
    Governorate:
      type: object
      required: [id, name, nameAr, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
    GovernorateInput:
      type: object
      required: [name, nameAr]
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
    GovernorateUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean

    # ─── AREA ──────────────────────────────────────────────────────────────────
    Area:
      type: object
      required: [id, name, nameAr, governorateId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        governorate:
          $ref: "#/components/schemas/Governorate"
        isActive:
          type: boolean
    AreaInput:
      type: object
      required: [name, nameAr, governorateId]
      properties:
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        isActive:
          type: boolean
    AreaUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean

    # ─── REQUEST ───────────────────────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, requestId, senderId, content, type, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
        type:
          type: string
          enum: [text, image]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content, type]
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image]
        imageUrl:
          type: string

    # ─── RATING ────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, title, body, type, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [new_request, new_offer, technician_selected, new_message, price_adjustment, status_change, support_reply, announcement]
        isRead:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ─── CMS ───────────────────────────────────────────────────────────────────
...